* All u/i database definitions must be int32 / uint32 , as LoadBlock(fields, Allocated); uses length as offset (offset += sizeof(uint32)). This solves rooted npc's, and probably more. Found wrong definitions in: Areatrigger, CreatureProto, CreatureNames and WorldInfo. Why it worked on some servers, and others not beats me.
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@899 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/Creature.h
@@ -61,8 +61,8 @@ struct CreatureInfo
 	uint32 Female_DisplayID2;
 	float unkfloat1;
 	float unkfloat2;
-	uint8  Civilian;
-	uint8  Leader;
+	uint32  Civilian;
+	uint32  Leader;
 
 	std::string lowercase_name;
 	GossipScript * gossip_script;
@@ -149,7 +149,7 @@ struct CreatureProto
 	uint32 extra_a9_flags;
 	uint32	auraimmune_flag;
 	int32 vehicle_entry;
-	uint8 CanMove;
+	uint32 CanMove;
 
 	/* AI Stuff */
 	bool m_canRangedAttack;
